requirements.txt
- Backend
Flask
Authlib
python-dotenv
requests
flask-cors
flask-uploads
flask-admin
flask-sqlalchemy
flask-login 
flask-mysqldb 
flask-session
pytest
pymysql
- Frontend
#@react-oauth/google